Fix: AI Prompt 파일 암호화 방식 변경 및 그에 따른 로직 수정
